Refactor client structure and add routing
Reorganized client-side code: moved CSS and asset files under src/assets, replaced UI renderers with React components (Form, Nav, Layout), and updated imports accordingly. Introduced react-router-dom and wired up BrowserRouter in App to render Form via a Layout with Outlet. Updated favicon path and removed legacy public icons.svg. Added a placeholder server/index.ts and updated package.json/package-lock to include react-router-dom and related deps.
