+++ client/src/assets/icons/hero.png
