+++ client/src/components/Layout.tsx
@@ -1,0 +1,18 @@
+import { Outlet } from "react-router";
+import Nav from "./Nav";
+
+import "../assets/css/layout.css";
+
+function Layout() {
+    return (
+        <div className="app-root">
+            <Nav />
+
+            <main className="main-content">
+                <Outlet />
+            </main>
+        </div>
+    );
+}
+
+export default Layout;
