+++ client/src/main.tsx
@@ -1,14 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import "./css/general.css";
+import "./assets/css/general.css";
 // import './index.css'
 
-import { App } from "./App.tsx";
-
-
-
-
-
+import App from "./App";
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
